apps: move volume related operations to a new file
